+++ components/FileFormats.vue
@@ -1,6 +1,6 @@
 <template>
 	<div class="vue-component file-formats">
-		<SearchableList :data="fileFormats" summaryKey="title" :showSummaryOnExpand="false" :externalSearchTerm="searchTerm" :sort="sort" :allowExpand="allowExpand" :heading="heading">
+		<SearchableList :data="fileFormats" summaryKey="title" :showSummaryOnExpand="false" :externalSearchTerm="searchTerm" :sort="sort" :offerDetails="offerDetails" :heading="heading" :collapsed="collapsed">
 			<template #heading="scope"><slot name="heading" v-bind="scope" /></template>
 			<template #summary="slot">
 				<strong class="inline">{{ slot.item.name }}</strong>
@@ -53,13 +53,17 @@ export default Utils.enableHtmlProps({
 			type: Boolean,
 			default: true
 		},
-		allowExpand: {
+		offerDetails: {
 			type: Boolean,
 			default: true
 		},
 		heading: {
 			type: String,
 			default: 'File Formats'
+		},
+		collapsed: {
+			type: Boolean,
+			default: false
 		}
 	},
 	computed: {
@@ -86,9 +90,6 @@ export default Utils.enableHtmlProps({
 			}
 			return data;
 		}
-	},
-	filters: {
-		abbrev: Utils.prettifyAbbreviation
 	}
 })
 </script>